refactor: consolidate RWA networks list and update imports
- Moved RWA_NETWORKS_LIST definition to useRwaTokens hook for better organization.
- Removed redundant imports of RWA_NETWORKS_LIST from various components.
- Updated tests to reflect the new structure and ensure proper functionality.
