+++ app/components/UI/Trending/hooks/useRwaTokens/useRwaTokens.ts
@@ -6,10 +6,27 @@ import {
   PriceChangeOption,
   SortDirection,
 } from '../../components/TrendingTokensBottomSheet';
-import { RWA_CHAIN_IDS } from '../../utils/trendingNetworksList';
+import { TRENDING_NETWORKS_LIST } from '../../utils/trendingNetworksList';
 import { NetworkToCaipChainId } from '../../../NetworkMultiSelector/NetworkMultiSelector.constants';
+import type { ProcessedNetwork } from '../../../../hooks/useNetworksByNamespace/useNetworksByNamespace';
 
 const RWA_PAGE_SIZE = 100;
+
+/**
+ * Networks supported for RWA (Real World Asset) tokens.
+ */
+export const RWA_NETWORKS_LIST: ProcessedNetwork[] =
+  TRENDING_NETWORKS_LIST.filter((n) =>
+    [
+      NetworkToCaipChainId.ETHEREUM,
+      NetworkToCaipChainId.BNB,
+      NetworkToCaipChainId.ROBINHOOD,
+    ].includes(n.caipChainId as NetworkToCaipChainId),
+  );
+
+export const RWA_CHAIN_IDS: CaipChainId[] = RWA_NETWORKS_LIST.map(
+  (n) => n.caipChainId,
+);
 
 /**
  * Default chains for the tokenized stocks feed (Ethereum + Robinhood).
