+++ app/components/UI/Trending/utils/trendingNetworksList.ts
@@ -4,7 +4,6 @@ import {
   TrxScope,
   ///: END:ONLY_INCLUDE_IF
 } from '@metamask/keyring-api';
-import type { CaipChainId } from '@metamask/utils';
 import { ProcessedNetwork } from '../../../hooks/useNetworksByNamespace/useNetworksByNamespace';
 import { getNetworkImageSource } from '../../../../util/networks';
 import { NetworkToCaipChainId } from '../../NetworkMultiSelector/NetworkMultiSelector.constants';
@@ -159,19 +158,3 @@ export const TRENDING_NETWORKS_LIST: ProcessedNetwork[] = [
     imageSource: getNetworkImageSource({ chainId: BtcScope.Mainnet }),
   },
 ];
-
-/**
- * Networks supported for RWA (Real World Asset) tokens.
- */
-export const RWA_NETWORKS_LIST: ProcessedNetwork[] =
-  TRENDING_NETWORKS_LIST.filter((n) =>
-    [
-      NetworkToCaipChainId.ETHEREUM,
-      NetworkToCaipChainId.BNB,
-      NetworkToCaipChainId.ROBINHOOD,
-    ].includes(n.caipChainId as NetworkToCaipChainId),
-  );
-
-export const RWA_CHAIN_IDS: CaipChainId[] = RWA_NETWORKS_LIST.map(
-  (n) => n.caipChainId,
-);
